finalize `LanguageStatusItem` API, https://github.com/microsoft/vscode/issues/129037
