+++ extensions/typescript-language-features/package.json
@@ -9,7 +9,6 @@
   "aiKey": "AIF-d9b70cd4-b9f9-4d70-929b-a071c400b217",
   "enabledApiProposals": [
     "inlayHints",
-    "languageStatus",
     "markdownStringBaseUri",
     "resolvers",
     "workspaceTrust"
