+++ extensions/vscode-api-tests/package.json
@@ -18,7 +18,6 @@
     "fsChunks",
     "inlayHints",
     "inlineCompletions",
-    "languageStatus",
     "markdownStringBaseUri",
     "notebookCellExecutionState",
     "notebookConcatTextDocument",
